+++ tests/testthat/test-api-pcate.R
@@ -1,3 +1,5 @@
+skip_if_missing_suggests("SuperLearner", "glmnet", "nprobust")
+
 set.seed(20051920) # 20051920 is derived from 'test'
 
 n <- 250